Level Down UI Functions, Building Party Combatants
UI in general still needs prettying up and more information available. However, the level down mechanics appear to be mostly working, even if battle damage is laughably too powerful at present. That's to be tweaked and messed with later when all abilities are, at least, functioning in a preliminary state.

Two of Jenny's Constructs (the sentry and the heal bot) are built and basically functioning. The sheild that is intended to help protect the party will be harder to figure out. Working out a 'Cover' ability where a Combatant takes the damage intended for other Combatants when active. Used for one of Domanic's abilities and Jenny's Shield Wall.